pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "webcalyzer"
version = "0.1.0"
description = "Telemetry extraction from launch webcast videos with interactive calibration."
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "OpenAI Codex" }
]
dependencies = [
  "matplotlib>=3.8",
  "numpy>=2.0",
  "opencv-python>=4.10",
  "pandas>=2.2",
  "pyyaml>=6.0",
  "rapidocr-onnxruntime>=1.4.4",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[project.scripts]
webcalyzer = "webcalyzer.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
